[CHORE] : 코드정리
