+++ app/src/main/java/com/sopt/now/compose/Profile/ProfileFriendItem.kt
@@ -1,6 +1,5 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.Profile
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.ui.Modifier
 
 import androidx.compose.foundation.layout.Row
@@ -8,38 +7,39 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.material3.SegmentedButtonDefaults.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.sopt.now.compose.Friend
+import com.sopt.now.compose.R
 
 
 @Composable
-fun ProfileUserItem(friend: Friend) {
+fun ProfileFriendItem(friend: Friend) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
-            .padding(vertical = 20.dp, horizontal = 10.dp)
-            .background(color = colorResource(id = R.color.skyblue))
-            .height(130.dp)
-        ,
+            .padding(vertical = 10.dp, horizontal = 10.dp)
+            .height(110.dp),
         verticalAlignment = Alignment.CenterVertically
     ){
-        val image: Painter = painterResource(id = R.drawable.profile)
+        val image: Painter = painterResource(id = R.drawable.ic_person_white_24)
         Image(
             painter = image,
-            contentDescription = "34기 YB 주효은입니다!",
-            modifier = Modifier.padding(start = 8.dp)
+            contentDescription = "Profile Image",
+            modifier = Modifier
+                .padding(start = 8.dp)
+                .padding(top = 8.dp)
+                .padding(bottom = 8.dp)
+
+
         )
         Spacer(modifier = Modifier.width(10.dp))
         Text(
@@ -58,7 +58,7 @@ fun ProfileUserItem(friend: Friend) {
 
 @Preview
 @Composable
-fun ProfileUserItemPreview() {
-    val friend = Friend(name = "hi", selfDescription = "sd", profileImage =R.drawable.profile )
+fun ProfileFriendItemPreview() {
+    val friend = Friend(name = "hi", selfDescription = "sd", profileImage = R.drawable.profile )
     ProfileFriendItem(friend = friend)
 }
